Movidos sistemas, prefabs y eventos a listas
