+++ bin/assets/scripts/Client/PrefabsList.lua
@@ -1,0 +1,5 @@
+--New Prefabs are to be required here for loading
+
+require('Bullet')
+require('Enemy')
+require('sampleScene-config')
